Fix initial policy mode in config - comma separated resources/operations
