+++ pkg/kubewarden/index.ts
@@ -5,19 +5,19 @@ import kubewardenRoutes from './routes/kubewarden-routes';
 import kubewardenStore from './store/kubewarden';
 
 // Init the package
-export default function(plugin: IPlugin) {
+export default function($plugin: IPlugin) {
   // Auto-import model, detail, edit from the folders
-  importTypes(plugin);
+  importTypes($plugin);
 
   // Provide plugin metadata from package.json
-  plugin.metadata = require('./package.json');
+  $plugin.metadata = require('./package.json');
 
   // Load product
-  plugin.addProduct(require('./config/kubewarden'));
+  $plugin.addProduct(require('./config/kubewarden'));
 
   // Add Vuex store
-  plugin.addDashboardStore(kubewardenStore.config.namespace, kubewardenStore.specifics, kubewardenStore.config);
+  $plugin.addDashboardStore(kubewardenStore.config.namespace, kubewardenStore.specifics, kubewardenStore.config);
 
   // Routes
-  plugin.addRoutes(kubewardenRoutes);
+  $plugin.addRoutes(kubewardenRoutes);
 }
