+++ pkg/kubewarden/plugins/policy-class.js
@@ -1,4 +1,4 @@
-import { colorForState, stateBackground, stateDisplay } from '@shell/plugins/dashboard-store/resource-class';
+import { colorForState, stateDisplay } from '@shell/plugins/dashboard-store/resource-class';
 import { get } from '@shell/utils/object';
 
 import KubewardenModel, { colorForStatus } from './kubewarden-class';
@@ -27,16 +27,6 @@ export default class PolicyModel extends KubewardenModel {
     }
 
     return stateDisplay();
-  }
-
-  get stateBackground() {
-    const color = this.colorForState;
-
-    if ( color ) {
-      return color.replace('text-', 'bg-');
-    }
-
-    return stateBackground();
   }
 
   get colorForState() {
